+++ firebase-vertexai/CHANGELOG.md
@@ -1,1 +1,4 @@
 # Unreleased
+* [feature] Added support for `responseMimeType` in `GenerationConfig`.
+* [changed] Renamed `GoogleGenerativeAIException` to `FirebaseVertexAIException`.
+